wip continued, better organized but not quite happy with it
